perc-api/requirements.txt
Flask
Flask-Cors
gunicorn
grpcio-tools
protoletariat
pytest
marshmallow